Created enum to control the color of every element in the game + create a script for the gems to change the color of the block when colliding
Light 'em Up:
In our ColorChanger script, create some logic so that a designer could choose between the three primary colors red, yellow, or blue in the inspector. In our Start() method, change the sprite's color to the selected color we've chosen for that game object. Create it in a way so that the designers can attach this script to any game object with a SpriteRenderer and easily choose the appropriate color in our ColorChanger script.

Magic Gem:
Create a new script we can attach to all our Gems in the Hierarchy. Create a way that our PlayerBlocks sprite color will change to the matching color of the gem when they come into contact with one another.