Spawn block with random color
Every press on the spawn button create a random int. This int is assigned to be one of the values in the color enum, and when the block is spawned, it spawn with a random color

Spawn A Random Colored Block:
Modify our SpawnNewBlock.cs script to instantiate a block prefab to one of our 3 random colors to a position of your choice.