Blocks gets destroy if their color is not the same as the spike
Added a check in the Hazard class to compare the color of the sprite renderer in both the hazard and the block

Spikey Hazard:
Modify the Hazard.cs spikes script to only destroy colored blocks that are not of the same color as the block itself. E.g., a blue block won't be destroyed by a blue spike.