Coin flash his color if the block is not of the same color
Flashy Coin:
Create a new script that we can attach to our Coin. Create a way so that if a PlayerBlock of the matching color runs over the coin, the coin is destroyed. If it's not a matching color, create a flashing animation modifying our sprites alpha transparency for visual feedback to the user.